+++ R/inferences_conformal.R
@@ -1,3 +1,5 @@
+# Shared utilities for conformal prediction methods
+
 get_conformal_score <- function(x, score, mfx = NULL) {
     if (is.null(mfx)) {
         mfx <- attr(x, "marginaleffects")
@@ -70,88 +72,4 @@ get_conformal_bounds <- function(x, score, conf_level, mfx = NULL) {
         x$pred.high <- x$estimate + d
     }
     return(x)
-}
-
-
-conformal_split <- function(x, test, calibration, score, conf_level, mfx = NULL, ...) {
-    # assertions
-    checkmate::assert_class(x, "predictions")
-    checkmate::assert_choice(
-        score,
-        choices = c("residual_abs", "residual_sq", "softmax")
-    )
-    checkmate::assert_data_frame(test, null.ok = FALSE)
-    checkmate::assert_data_frame(calibration, null.ok = FALSE)
-
-    # calibration
-    # use original model---fitted on the training set---to make predictions in the calibration set
-    # p_calib is the `predictions()` call, which we re-evaluate on newdata=calibration
-    p_calib <- mfx@call
-    p_calib[["newdata"]] <- calibration
-    p_calib[["vcov"]] <- FALSE # faster
-    p_calib <- eval(p_calib)
-    score <- get_conformal_score(p_calib, score = score, mfx = mfx)
-
-    # test
-    # use original model to make predictions in the test set
-    p_test <- mfx@call
-    p_test[["newdata"]] <- test
-    p_test <- eval(p_test)
-
-    # bounds
-    out <- get_conformal_bounds(p_test, score = score, conf_level = conf_level, mfx = mfx)
-
-    return(out)
-}
-
-
-conformal_cv_plus <- function(x, test, R, score, conf_level, mfx = NULL, ...) {
-    # assertions
-    checkmate::assert_class(x, "predictions")
-    checkmate::assert_choice(
-        score,
-        choices = c("residual_abs", "residual_sq", "softmax")
-    )
-    checkmate::assert_data_frame(test, null.ok = FALSE)
-    checkmate::assert_integerish(R, upper = 25)
-
-    # cross-validation
-    train <- mfx@modeldata
-    idx <- sample.int(nrow(train), nrow(train))
-    idx <- split(idx, ceiling(seq_along(idx) / (length(idx) / R)))
-    scores <- NULL
-    for (i in idx) {
-        data_cv <- train[-i, ]
-        # re-fit the original model on training sets withholding the CV fold
-        model_cv <- tryCatch(stats::update(mfx@model, data = data_cv),
-            error = function(e) NULL)
-        if (is.null(model_cv)) {
-            if (is.call(mfx@call_model) && "data" %in% names(mfx@call_model)) {
-                # if the model call has a data argument, we can update it
-                mfx@call_model$data <- data_cv
-                model_cv <- eval(mfx@call_model)
-            } else {
-                stop_sprintf("Failed to re-fit the model on the cross-validation set.")
-            }
-        }
-        # use the updated model to make out-of-fold predictions
-        # call_cv is the `predictions()` call, which we re-evaluate in-fold: newdata=train[i,]
-        call_cv <- mfx@call
-        call_cv[["model"]] <- model_cv
-        call_cv[["newdata"]] <- train[i, ]
-        call_cv[["vcov"]] <- FALSE # faster
-        pred_cv <- eval(call_cv)
-        # save the scores form each fold
-        scores <- c(scores, get_conformal_score(pred_cv, score = score, mfx = mfx))
-    }
-
-    # test
-    out <- mfx@call
-    out[["newdata"]] <- test
-    out <- eval(out)
-
-    # bounds
-    out <- get_conformal_bounds(out, score = scores, conf_level = conf_level, mfx = mfx)
-
-    return(out)
 }